LALALALA I DELETED EVERYTHING!! :D
Kidding.  Deleted all the extraneous stuff that *wasn't* in the DependencyVisualizer directory because I didn't need it.  Moved all the stuff that *was* in that directory to root because I did need it.  Plan now is to turn this into a node module.  Let's see how this goes!

P.S. This is probably going to F up my git history beyond belief.